Fix tryFinishTrack

+++ java/need-for-speed/src/main/java/NeedForSpeed.java
@@ -62,7 +62,7 @@ class RaceTrack {
 	}
 
 	public boolean tryFinishTrack(NeedForSpeed car) {
-		if (((distance / car.getSpeed()) * car.getBatteryDrain()) > 100) {
+		if ((((double) distance / car.getSpeed()) * car.getBatteryDrain()) > 100) {
 			return false;
 		} else {
 			return true;
